Extract derivative functionality.
This avoids cluttering the interface for assets with derivative logic, while
simultaneously requiring that "derivative creation" be explicitly requested when
it's needed. This keeps the dependency graph pointing in one direction.
