+++ app/decorators/document_with_derivatives.rb
@@ -1,0 +1,11 @@
+class DocumentWithDerivatives < AssetWithDerivatives
+  def runners
+    [
+      injector.pdf_runner(id)
+    ]
+  end
+
+  def derivative_class
+    OregonDigital::Derivatives::Generators::DocumentDerivativeGenerator
+  end
+end
