+++ app/models/generic_asset.rb
@@ -5,12 +5,12 @@ class GenericAsset < ActiveFedora::Base
   contains "content", :class_name => 'FileContent'
   contains "workflow_metadata", :class_name => "Files::YmlFile"
 
-  private
-
   def injector
     @injector ||= OregonDigital.inject
   end
-  
+
+  private
+
   def assign_id
     OregonDigital::IdService.mint
   end
